Align behaviour of hook-sysconfig with hook-distutils and refactor code.
Change hook-sysconfig to use same method as hook-distutils to get
relative path to data files, and refactor the duplicate code into
hookutils.

This also prevents hook-sysconfig from causing a "security error"
because it used the wrong method of getting the containing folder.
